requirements.txt
djangorestframework
psycopg2-binary
openpyxl
pandas
djangorestframework-simplejwt